version3

+++ main.js
@@ -1,14 +1,25 @@
+var score = 0;
+var scoreString = '';
+var scoreText;
+var lives;
+var stateText;
+
 var main = {
     preload: function() {
         // Loat game sprites.
         game.load.image('background', 'assets/background.png');
         game.load.image('pixel', 'assets/pixel.png');
-        game.load.image('bullet', 'assets/bullet.png');
+        game.load.image('live', 'assets/live.png');
         /// Load block spritesheet.
         game.load.spritesheet('obstacle', 'assets/obstacle.png', 46, 100);
         game.load.spritesheet('enemy', 'assets/enemy.png', 53, 86);
         game.load.spritesheet('explosion', 'assets/explosion.png', 128, 128);
         game.load.spritesheet('player', 'assets/player.png', 144, 128);
+        game.load.spritesheet('bullet', 'assets/flame.png', 103, 103);
+        game.load.spritesheet('bullet2', 'assets/bomb1.png', 282, 234);
+        game.load.spritesheet('coin', 'assets/coin.png', 46, 100);
+
+        
     },
     create: function() {
         game.physics.startSystem(Phaser.Physics.ARCADE);
@@ -29,13 +40,13 @@ var main = {
         ///enemy
         this.enemyPool = game.add.group();
         this.enemyPool.enableBody = true;
-        this.enemyPool.createMultiple(50, 'enemy');
+        this.enemyPool.createMultiple(100, 'enemy');
         this.enemyPool.setAll('anchor.x', 0.5);
         this.enemyPool.setAll('anchor.y', 0.5);
         this.enemyPool.setAll('outOfBoundsKill', true);
         this.enemyPool.setAll('checkWorldBounds', true);
         this.enemyPool.forEach(function(enemy) {
-            enemy.animations.add('enemy_fly',[0, 1, 2, 3, 4], 20, true);
+            enemy.animations.add('enemy_fly',[0, 1, 2, 3, 4], 15, true);
         });
         this.nextEnemyAt = 0;
         this.enemyDelay = 1500;
@@ -48,6 +59,9 @@ var main = {
         this.bulletPool.setAll('anchor.y', 1);
         this.bulletPool.setAll('outOfBoundsKill', true);
         this.bulletPool.setAll('checkWorldBounds', true);
+        this.bulletPool.forEach(function(bullet) {
+            bullet.animations.add('bullet');
+        });
         this.nextShotAt = 0;
         this.shotDelay = 100;   
         
@@ -63,10 +77,31 @@ var main = {
             obstacle.animations.add('obstacle_fly',[0, 1, 2, 3, 4], 10, true);
         });
         this.nextobstacleAt = 0;
-        this.obstacleDelay = 50000;
+        this.obstacleDelay = 15000;
 
         ///control
         this.cursors = game.input.keyboard.createCursorKeys();
+
+        ///score
+        scoreString = 'Score : ';
+        scoreText = game.add.text(10, 10, scoreString + score, { font: '34px Georgia', fill: '#000' });
+
+        ///lives
+        lives = game.add.group();
+        game.add.text(game.world.width - 100, 10, 'Lives : ', { font: '34px Georgia', fill: '#000' });
+
+        ///Text
+        stateText = game.add.text(game.world.centerX,game.world.centerY,' ', { font: '84px Georgia', fill: '#000' });
+        stateText.anchor.setTo(0.5, 0.5);
+        stateText.visible = false;
+
+        for (var i = 0; i < 3; i++) 
+        {
+            var live = lives.create(game.world.width - 100 + (30 * i), 60, 'live');
+            live.scale.setTo(0.008, 0.008);
+            // live.alpha = 0.4;
+        }
+
         /// Particle
         // this.emitter = game.add.emitter(422, 320, 150);
         // this.emitter.makeParticles('pixel');
@@ -76,7 +111,21 @@ var main = {
         // this.emitter.gravity = 0;
     },
 
-
+    update: function() {
+        ///  Scroll the background
+        this.bg.tilePosition.y += 2;        
+        ///enemy
+        this.GenerateEnemy();
+        ///obstacle
+        this.GenerateObstacle();
+        ///player
+        this.MovePlayer();
+        ///collide
+        game.physics.arcade.overlap(this.player, this.enemyPool, this.playerHit, null, this);
+        game.physics.arcade.overlap(this.bulletPool, this.enemyPool, this.enemyHit, null, this);
+        game.physics.arcade.overlap(this.enemyPool, this.obstaclePool, this.obstacleHit, null, this);
+        game.physics.arcade.collide(this.player, this.obstaclePool, null, null, this);
+    }, 
     explosion: function() {
 
         /// ToDo 5: Start our emitter.
@@ -85,26 +134,8 @@ var main = {
         ///      3. Set frequency to null since we will burst out all partice at once.
         ///      4. We'll launch 15 particle.
 
-        ///
+
     },
-    update: function() {
-        ///sea
-        //  Scroll the background
-        this.bg.tilePosition.y += 2;        
-        ///collide
-        game.physics.arcade.overlap(this.player, this.enemyPool, this.playerHit, null, this);
-        game.physics.arcade.overlap(this.bulletPool, this.enemyPool, this.enemyHit, null, this);
-        // game.physics.arcade.collide(this.enemy, this.obstaclePool, null, null, this);
-        game.physics.arcade.collide(this.player, this.obstaclePool, null, null, this);
-
-        //enemy
-        this.GenerateEnemy();
-
-        this.GenerateObstacle();
-
-
-        this.MovePlayer();
-    }, 
     playerDie: function() { game.state.start('main');},
     GenerateObstacle: function(){
         if (this.nextobstacleAt<game.time.now && this.obstaclePool.countDead()>0) {
@@ -112,6 +143,8 @@ var main = {
             var obstacle = this.obstaclePool.getFirstExists(false);
             obstacle.reset(game.rnd.integerInRange(20, 780), 0);
             obstacle.body.velocity.y = game.rnd.integerInRange(30, 60);
+            // obstacle.reset(0,game.rnd.integerInRange(20, 580));
+            // obstacle.body.velocity.x = game.rnd.integerInRange(30, 60);
             obstacle.play('obstacle_fly');
         }
     },
@@ -121,6 +154,8 @@ var main = {
             var enemy = this.enemyPool.getFirstExists(false);
             enemy.reset(game.rnd.integerInRange(20, 780), 0);
             enemy.body.velocity.y = game.rnd.integerInRange(30, 60);
+            enemy.body.velocity.x = game.rnd.integerInRange(30, 60);
+
             enemy.play('enemy_fly');
         }
     },
@@ -149,14 +184,40 @@ var main = {
         
  
     },
-    obstacleHit: function(player, obstacle) { 
-        // var explosion = game.add.sprite(player.x, player.y, 'explosion');
-        // explosion.anchor.setTo(0.5);
-        // explosion.animations.add('boom');
-        // explosion.play('boom', 15, false, true);
+    obstacleHit: function(enemy, obstacle) { 
+        ///Increase the score
+        score += 50;
+        scoreText.text = scoreString + score;
+        if (score>=1000)
+        {
+            score += 1000;
+            scoreText.text = scoreString + score;
 
+            stateText.text = " You Won, \n Click to restart";
+            stateText.visible = true;
+        }
+        ///
+        enemy.animations.add('enemy_hit',[5, 6], 20, true);
+        enemy.play('enemy_hit');
+        var explosion = game.add.sprite(enemy.x, enemy.y, 'explosion');
+        explosion.anchor.setTo(0.5);
+        explosion.animations.add('boom');
+        explosion.play('boom', 15, false, true);
+        enemy.kill();;
+        game.load.spritesheet('bullet', 'assets/bomb1.png', 282, 234);
     },
     enemyHit: function(bullet, enemy) {
+        ///Increase the score
+        score += 10;
+        scoreText.text = scoreString + score;
+        if (score>=1000)
+        {
+            score += 1000;
+            scoreText.text = scoreString + score;
+            stateText.text = " You Won, \n Click to restart";
+            stateText.visible = true;
+        }
+        ///
         enemy.animations.add('enemy_hit',[5, 6], 20, true);
         enemy.play('enemy_hit');
         bullet.kill();
@@ -173,7 +234,19 @@ var main = {
         explosion.anchor.setTo(0.5);
         explosion.animations.add('boom');
         explosion.play('boom', 15, false, true);
-        player.kill();
+
+        ///live
+        live = lives.getFirstAlive();
+        if (live)
+        {
+            live.kill();
+        }
+        if (lives.countLiving() < 1)
+        {
+            player.kill();
+            stateText.text=" GAME OVER \n Click to restart";
+            stateText.visible = true;
+        }
     },
     fire: function() { 
         if (!this.player.alive || this.nextShotAt>game.time.now)
@@ -184,7 +257,9 @@ var main = {
         var bullet = this.bulletPool.getFirstExists(false);
         bullet.reset(this.player.x, this.player.y-20);
         bullet.body.velocity.y = -500;
+        bullet.play('bullet',true,true);
     },
+
 
 };
